ol_osd_module: fix interactions between set_lrc() and set_message()
When ol_osd_module_set_lrc() is called, it tries to clear the current
message (if any). Unfortunately, this doesn't work because
hide_message() does nothing if osd->lrc is set. This causes two
problems:

1. The current message is not cleared. This is usually unnoticeable,
   because the lyrics are updated every 100 ms while playing, but if the
   player is currently stopped, then the message will just stay there.

2. In these circumstances, hide_message() omits to clear
   osd->message_source, leading to other problems later on (code
   believing hide_message() is scheduled later on, when it's not).

To fix both problems, modify set_lrc() so that it sets osd->lrc *after*
clearing the current message.

To make sure that hide_message() is never called with osd->lrc set, also
remove the clearing of lyrics/message from _update_metadata(), which
is not needed anyway (the "track-changed" signal is handled by ol_main,
which calls the display modules' set_lrc(), clearing the lyrics/message
as appropriate).