ol_osd_module: always reset lyrics state in set_lrc()
If set_lrc() is called with a non-null lrc_file and we are currently
displaying lyrics (non-null priv->lrc), we need to reset the internal
lyrics state. However we don't need to clear the display in that case,
that's deferred to set_played_time().

Therefore, split clear_lyrics() into reset_lyrics_state() and
hide_lyrics() (display only). Always call the former in set_lrc(), and
call the latter only if the new lrc_file is null.