ol_osd_module: force refreshing on first set_played_time()
When set_lrc() is called with a non-null lrc_file, we don't clear the
display. This is normally fine, as the display will be refreshed in
set_played_time(). However, this doesn't work when lrc_file is empty,
because set_played_time() will not find any lyrics to display and will
just bail out.

To work around this, set a new force_refresh_on_set_played_time boolean
in set_lrc() when we don't clear the display, and always clear the
display in set_played_time() if it is set. The boolean is then cleared,
because we only want to force the clearing on the first
set_played_time().