ol_osd: fix show/hide glitch
Not sure what is going here _exactly_, but here goes what I've gathered.
I've noticed that when hiding the OSD (e.g. by clicking on the tray
icon), and then showing it again, it often ends up being repositioned to
(0,0), i.e. its position is lost.

After much pointless debugging, I got lucky and found out that this is
actually caused by the OSD toolbar, in the following scenario:
1. Start OSD Lyrics, OSD shown.
2. Start playing.
3. Hide the OSD.
4. Stop playing (and possibly start again).
5. Show the OSD.

Indeed, the toolbar listens to the "status-changed" signal, and changes
the visibility of the play/pause buttons accordingly. For some reason,
hiding a button while the toolbar is not visible (because the OSD window
is hidden) is causing the window to be moved to (0,0), which only
becomes apparent once the window is shown again.

Not sure why this is happening really. Regardless, this commit is
working around this by informing the toolbar when the window's
visibility changes, so that no button gets hidden while the window
itself is hidden. When the window is shown again, the buttons'
visibility is explicitly updated.

While at it, also remove the gtk_widget_queue_draw() from
_update_status(), I don't see why this would be needed.