Rename config entry ".visible" -> "OSD/visible"
The config entry that controls the visibility of the OSD was originally
named "OSD/visible" (before commit 638c86b60d92 ("implement default
value setting")). Through some accident of history, it ended up being
called either ".visible" or "General/visible", resulting in the keyboard
shortcut to toggle the visibility doing nothing.

The motivation for using ".visible" was to make the option
non-persistant (see src/ol_trayicon.c). However this has the potentially
undesirable side effect of making it impossible to set it through DBus.
"General/visible" seems to be a strange name as well, as it only
toggles the visibility of the OSD.

All things considered, it feels like reverting to the original name is
the most sensible thing to do.

Fixes #49