ol_scroll_module: do not allow the close button to grab focus
For a while I could not explain why pressing Space or Enter in the
scroll window causes OSD Lyrics to exit. It turned out this is because
the close button always gets the focus when the window is active, causing
Space/Enter to "click" it.

This is clearly not desirable, so make sure that the close button cannot
grab the focus (i.e. it can only be activated by actually clicking on it).