ol_osd: hide the OSD only while the player is stopped
Right now we have two rather redundant mechanisms:
- Switching the OSD window (creating/destroying it)
- Hiding the OSD window

There isn't really a good reason to have both, unless you are hiding the
OSD temporarily quite often (faster than creating/destroying the
window). From experience this is typically because you are not playing
anything and you want the OSD window out of the way.

Well, maybe the best thing is simply not to show the OSD window when the
player is stopped, and this way it makes sense to have two mechanisms :)

This commit is really doing two things:
1. OSD/visible becomes OSD/visible_when_stopped and its meaning changes
   accordingly. To achieve that behaviour, the "status-changed" callback
   is moved from ol_osd_toolbar to ol_osd_module, and the module then
   feeds status updates to the toolbar.
2. The shortcuts to hide the OSD (Ctrl-Shift-H and clicking on the tray
   icon) now trigger the OSD switching mechanism
   (General/display-mode-osd) instead. This way they broadly do the same
   thing as before. There is clearly no point in having shortcuts for
   OSD/visible_when_stopped, as it is something you probably never
   change once configured.

Note: this partially reverts commit ece28b2a3079 ("ol_osd: fix show/hide
glitch"), indeed a nice side-effect of driving the toolbar status
updates from the module is that we no longer need to deal with the bug
this commit was trying to fix. Also, fun fact,
ol_osd_toolbar_set_status() is not added to the header as it is already
there! That function was introduced a long time ago, then removed, but
stayed in the header. So a bit more partial reverting there.