players: Fix position update when seeking
The way the position is updated when we get a "Seeked" signal from an
MPRIS2 player is rather convoluted: we discard the position provided by
the signal, and instead update the position by asking the player again.

This is clearly suboptimal (unnecessary D-Bus call), but it turns out to
be unreliable as well: it is possible that the player will return a
stale position if asked just after seeking. As a result, OSD Lyrics
sometimes gets (completely) out of sync when the player is seeked. I've
seen this issue with gmusicbrowser but I may well affect other players
too.

This commit does the obvious: instead of asking the player again, just
forward the position provided by the signal to position_changed().
HttpPlayer is also changed accordingly (and the timer code removed as it
is redundant with position_changed()).