ol_lyric_source: try to infer artist from title if invalid
When searching for lyrics, try to do some basic sanitization of the
artist/title fields obtained from the player. Specifically, if the
artist field is empty or invalid, try to parse the title field in case
it actually contains multiple tags.

This is based on and replaces #80 by @elementdavv, aiming to be more
careful (not making too many assumptions and does nothing if the artist
field is valid) and faster (the strings are only duplicated if needed,
both the title and artist are generated at the same time).

Note there is a small difference compared to #80: <artist>--<title> is
supported instead of <title>--<artist>, as the latter seems like a
strange format.